control flow for filling in glo30 with glo90
